requirements.txt
tensorflow-cpu>=2.17.0
joblib
numpy
groq>=0.9.0
pillow
scikit-learn
duckduckgo-search
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
langchain
langchain-community
